+++ tests/src/frontend/commands/BlockControllerTest.php
@@ -324,9 +324,10 @@ EOT;
  * @param $this->varValue('foo');
  * @param $this->varValue('foo');
  *
- * @var $this \luya\cms\base\PhpBlockView
+ * @var \luya\cms\base\PhpBlockView $this
  */
 ?>
+
 EOT;
         $this->assertSame($view, $ctrl->generateViewFile('MySuperBlock'));
     }