+++ CHANGELOG.md
@@ -21,6 +21,7 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 + [#194](https://github.com/luyadev/luya-module-cms/issues/194) Updated CSS to unify spaces between page languages and treenav
 + [#71](https://github.com/luyadev/luya-module-cms/issues/71) If a not found page is defined a 404 is returned instead of redirect to the error page.
 + [#20](https://github.com/luyadev/luya-module-cms/issues/20) Refactor parent page selection and menu-dropdown directive (added search, change styles).
++ [#189](https://github.com/luyadev/luya-module-cms/issues/189) Block preview path with backward compatibility for PHP 5.6.
 
 ## 1.0.9 (18. February 2019)