+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 2.0.1
+
+### Fixed
+
++ [#209](https://github.com/luyadev/luya-module-cms/pull/209) Fixed phpdoc in view file generator of blocks.
+
 ## 2.0.0 (26. June 2019)
 
 > This release contains new migrations and requires to run the migrate command after updating. Check the [UPGRADE document](UPGRADE.md) to read more about breaking changes.